+++ src/objimpl.rs
@@ -55,7 +55,7 @@ pub unsafe fn PyType_SUPPORTS_WEAKREFS(t : *mut PyTypeObject) -> c_int {
 
 #[inline(always)]
 pub unsafe fn PyObject_GET_WEAKREFS_LISTPTR(o : *mut PyObject) -> *mut *mut PyObject {
-    let weaklistoffset = (*Py_TYPE(o)).tp_weaklistoffset as int;
+    let weaklistoffset = (*Py_TYPE(o)).tp_weaklistoffset as isize;
     (o as *mut c_char).offset(weaklistoffset) as *mut *mut PyObject
 }
